+++ common/persistence/serialization/serializer_test.go
@@ -167,10 +167,6 @@ func (s *temporalSerializerSuite) TestSerializeShardInfo_EmptyMapSlice() {
 	shardInfo.RangeId = rand.Int63()
 
 	categoryID := rand.Int31()
-	shardInfo.QueueAckLevels = make(map[int32]*persistencespb.QueueAckLevel)
-	shardInfo.QueueAckLevels[categoryID] = &persistencespb.QueueAckLevel{
-		ClusterAckLevel: make(map[string]int64),
-	}
 	shardInfo.QueueStates = make(map[int32]*persistencespb.QueueState)
 	shardInfo.QueueStates[categoryID] = &persistencespb.QueueState{
 		ReaderStates: make(map[int64]*persistencespb.QueueReaderState),
