+++ proto/internal/temporal/server/api/persistence/v1/queues.proto
@@ -26,11 +26,6 @@ option go_package = "go.temporal.io/server/api/persistence/v1;persistence";
 import "temporal/server/api/persistence/v1/predicates.proto";
 import "temporal/server/api/persistence/v1/tasks.proto";
 
-message QueueAckLevel {
-    int64 ack_level = 1;
-    map<string, int64> cluster_ack_level = 2;
-}
-
 message QueueState {
     map<int64, QueueReaderState> reader_states = 1;
     TaskKey exclusive_reader_high_watermark = 2;
